+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "log-collection-backend"
+name = "log-collection-client"
 version = "0.1.0"
 edition = "2021"
