+++ client.py
@@ -1,10 +1,16 @@
 import socket
+import sys
+
+if (len(sys.argv) < 2):
+    print('Usage: python3 client.py "message"');
+    sys.exit(1)
+
 # Target IP address and port
 target_ip = '52.15.167.81'  # Replace with the IP address of the target machine
 target_port = 45612  # Replace with the desired port number
 
 # Data to be sent
-data_to_send = "data sending from local socket to central test server"
+data_to_send = sys.argv[1];
 
 # Create a socket
 client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
